Impl M2M binary update ...
* Use getopt in proone-bne
* Call pth_yield() where necessary
* htbt: truncate downloaded binary if actual size of data transfered is
  less than alloc_len
* prne_rcb_param.self is not optional as proone-bne uses nybin
* Add --vercmp option for proone-bne to test M2M bin update
* Refactor proone-htbtclient
* Protocol change: prne_htbt_bin_meta.alloc_len is now size_t
* Add convenience functions: prne_realloc_str(), prne_redup_str()

+++ src/protocol.h
@@ -278,7 +278,7 @@ struct prne_htbt_cmd {
 };
 
 struct prne_htbt_bin_meta {
-	uint32_t alloc_len;
+	size_t alloc_len;
 	prne_htbt_cmd_t cmd;
 };
